requirements.txt
Django==2.2.12
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
